Clean-up: Changing routes from playground to characters, de-nesting controllers from routes and other minimal ajustments through out the structure
