+++ index.js
@@ -4,8 +4,7 @@ import express from "express";
 import cookieParser from "cookie-parser";
 import fs from "fs";
 import errorHandler from "./middlewares/errorHandler.js"
-import playgroundRouter from "./routes/playgroundRouter.js";
-// import userTrackingRequests from "./middlewares/userTrackingRequests.js";
+import playgroundRouter from "./routes/characterRouter.js";
 import tracker from "./middlewares/tracker.js";
 import { nextTick } from "process";
 
@@ -16,17 +15,6 @@ app.use(cookieParser());
 app.use(express.json());   // middleware that processes the incoming raw json
 app.use(tracker);
 app.use('/playground', playgroundRouter);
-app.get('/', (req, res, next) => {         
-    console.log(req.cookies);
-    fs.readFile('/file-does-not-exist', (err, data) => { // when we have asyncronous code, call-backs or promisses in express, we need to call next
-        if(err) {
-            next(err);
-        } else {
-            res.send(data);
-        }
-    });
-      
-});                                        
 
 app.use(errorHandler);    // the order matters here; it has to be at the end
