+++ middlewares/asyncHandler.js
@@ -1,0 +1,8 @@
+function asyncHandler(fn) {
+    return function(req, res, next){
+        return Promise.resolve(fn(req, res, next)).catch(next);
+    };
+}
+
+
+export default asyncHandler;
