--- middlewares/userTrackingRequests.js
@@ -1,14 +1,0 @@
-/*export const userTrackingRequests = (req, res, next) => {         //It's not working , why??????
-    const {
-        body: { id, Name }
-    } = req;
-    if (id && Name) {
-        req.id = id;
-        req. Name = Name;
-        next();
-    } else {
-        res.sendStatus(401);
-    }
-};
-
-export default userTrackingRequests; */
