+++ utils/ErrorResponse.js
@@ -1,0 +1,8 @@
+class ErrorResponse extends Error {
+    constructor(message, statusCode = 500) {
+        super(message);
+        this.statusCode = statusCode;
+    }
+}
+
+export default ErrorResponse;
